Move tests into namespaces
