+++ Toxon.UrlTemplates.UnitTests/Parser/ParserTests.Level1Expressions.cs
@@ -1,14 +1,13 @@
-﻿using System.Linq;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
-namespace Toxon.UrlTemplates.UnitTests
+namespace Toxon.UrlTemplates.UnitTests.Parser
 {
     public partial class ParserTests
     {
         [Test]
-        public void Level3Variable()
+        public void Level1Variable()
         {
-            var input = "{.foo123_%20}";
+            var input = "{foo123_%20}";
             var result = Parse(input);
 
             AssertOnSuccess(result, x =>
@@ -19,9 +18,9 @@ namespace Toxon.UrlTemplates.UnitTests
             });
         }
         [Test]
-        public void Level3MultiVariable()
+        public void Level1LiteralVariableLiteral()
         {
-            var input = "{.foo123_%20,bar456_%21}";
+            var input = "hello{foo}world";
             var result = Parse(input);
 
             AssertOnSuccess(result, x =>
