+++ Toxon.UrlTemplates.UnitTests/Parser/ParserTests.Level4Expressions.cs
@@ -1,7 +1,6 @@
-﻿using System.Linq;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
-namespace Toxon.UrlTemplates.UnitTests
+namespace Toxon.UrlTemplates.UnitTests.Parser
 {
     public partial class ParserTests
     {
