+++ Toxon.UrlTemplates.UnitTests/Toxon.UrlTemplates.UnitTests.csproj
@@ -41,15 +41,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ExecutorTests.cs" />
+    <Compile Include="Execution\ExecutorTests.cs" />
     <Compile Include="ExtendedTests.cs" />
-    <Compile Include="ParserTests.cs" />
-    <Compile Include="ParserTests.Level4Expressions.cs" />
-    <Compile Include="ParserTests.Level3Expressions.cs" />
-    <Compile Include="ParserTests.Level2Expressions.cs" />
-    <Compile Include="ParserTests.Level1Expressions.cs" />
-    <Compile Include="ParserTests.Literals.cs" />
-    <Compile Include="PartialExpansionTests.cs" />
+    <Compile Include="Parser\ParserTests.cs" />
+    <Compile Include="Parser\ParserTests.Level4Expressions.cs" />
+    <Compile Include="Parser\ParserTests.Level3Expressions.cs" />
+    <Compile Include="Parser\ParserTests.Level2Expressions.cs" />
+    <Compile Include="Parser\ParserTests.Level1Expressions.cs" />
+    <Compile Include="Parser\ParserTests.Literals.cs" />
+    <Compile Include="Execution\PartialExpansionTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReferenceTests.cs" />
     <Compile Include="RfcBySectionTests.cs" />
